feat: separating tables definition in another file
